refactor: remove src folder
